Default setName in DisplayObject constructors

+++ src/main/java/com/ancevt/d2d2/display/IDisplayObject.java
@@ -21,7 +21,7 @@ import com.ancevt.d2d2.event.IEventDispatcher;
 
 public interface IDisplayObject extends IEventDispatcher {
 
-    long displayObjectId();
+    int displayObjectId();
 
     String getName();
